validation and exceptions added

+++ service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -4,11 +4,7 @@ import com.epam.esm.converters.DtoToTagConverter;
 import com.epam.esm.converters.TagToDtoConverter;
 import com.epam.esm.dao.TagDao;
 import com.epam.esm.dto.TagDto;
-import com.epam.esm.exceptions.EntityAlreadyExistsException;
-import com.epam.esm.exceptions.EntityNotFoundException;
-import com.epam.esm.exceptions.InternalServerException;
-import com.epam.esm.exceptions.InvalidDataProvidedException;
-import com.epam.esm.model.Tag;
+import com.epam.esm.exceptions.*;
 import com.epam.esm.service.TagService;
 import com.epam.esm.validators.GiftCertificateValidator;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 
 @Service
 public class TagServiceImpl implements TagService {
@@ -38,13 +34,16 @@ public class TagServiceImpl implements TagService {
     @Override
     @Transactional
     public TagDto create(TagDto tagDto) {
+        if(Objects.isNull(tagDto)){
+            throw new EntityCreationException();
+        }
         if(validator.isValidTagName(tagDto.getName())){
             if (tagDao.findByName(tagDto.getName()).isEmpty()){
                 return tagToDtoConverter.convert(tagDao.insert(dtoToTagConverter.convert(tagDto)));
             }
             throw new EntityAlreadyExistsException();
         }
-        throw new InvalidDataProvidedException();
+        throw new InvalidDataProvidedException("invalid tag name - " + tagDto.getName());
     }
 
     @Override
@@ -55,29 +54,38 @@ public class TagServiceImpl implements TagService {
     @Override
     @Transactional
     public TagDto findById(Long id) {
-        return tagDao.findById(id).map(tagToDtoConverter::convert).orElseThrow(EntityNotFoundException::new);
+        if(Objects.nonNull(id)){
+            return tagDao.findById(id).map(tagToDtoConverter::convert).orElseThrow(EntityNotFoundException::new);
+        }
+        throw new EntityNotFoundException();
     }
 
     @Override
     @Transactional
     public TagDto update(Long id, TagDto tagDto) {
+        if(Objects.isNull(id) || Objects.isNull(tagDto)){
+            throw new EntityModifyingException();
+        }
         TagDto foundTag = findById(id);
         if(validator.isValidTagName(tagDto.getName())){
             if (tagDao.update(id, dtoToTagConverter.convert(tagDto)) > 0){
                 foundTag.setName(tagDto.getName());
                 return foundTag;
             }
-            throw new InternalServerException();
+            throw new EntityModifyingException();
         }
-        throw new InvalidDataProvidedException();
+        throw new InvalidDataProvidedException("invalid tag name - " + tagDto.getName());
     }
 
     @Override
+    @Transactional
     public TagDto delete(Long id) {
-        TagDto tag = findById(id);
-        if(tagDao.deleteById(id) > 0){
-            return tag;
+        if(Objects.nonNull(id)){
+            TagDto tag = findById(id);
+            if(tagDao.deleteById(id) > 0){
+                return tag;
+            }
         }
-        throw new EntityNotFoundException();
+        throw new EntityModifyingException();
     }
 }
